% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdds.R
\name{Normal}
\alias{Normal}
\title{Generate random deviates from a normal distribution}
\description{
Generate random deviates from a normal distribution

Generate random deviates from a normal distribution
}
\examples{

## ------------------------------------------------
## Method `Normal$new`
## ------------------------------------------------

\dontrun{
  sc <- sparklyr::spark_connect(master = "local")
  norm <- Normal$new(
    sc = sc,
    size = 10L,
    num_partitions = 1L,
    seed = 1L
  )
}

## ------------------------------------------------
## Method `Normal$collect`
## ------------------------------------------------

\dontrun{
  norm$collect()
}

## ------------------------------------------------
## Method `Normal$count`
## ------------------------------------------------

\dontrun{
  norm$collect()
}

## ------------------------------------------------
## Method `Normal$first`
## ------------------------------------------------

\dontrun{
  norm$count()
}

## ------------------------------------------------
## Method `Normal$get_num_partitions`
## ------------------------------------------------

\dontrun{
  norm$get_num_partitions()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Normal$new()}}
\item \href{#method-collect}{\code{Normal$collect()}}
\item \href{#method-count}{\code{Normal$count()}}
\item \href{#method-first}{\code{Normal$first()}}
\item \href{#method-get_num_partitions}{\code{Normal$get_num_partitions()}}
\item \href{#method-clone}{\code{Normal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Normal} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$new(sc, size, num_partitions, seed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc}}{A \code{spark_connection}.}

\item{\code{size}}{\code{integer(1)}. The number of deviates to generate.}

\item{\code{num_partitions}}{\code{integer(1)}. Number of partitions in the RDD
(default: \code{sc.defaultParallelism}).}

\item{\code{seed}}{\code{integer(1)}. The seed to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Normal} object which is a \code{RDD[Double]} comprised of
comprised of i.i.d. samples ~ \verb{N(0.0, 1.0).}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  sc <- sparklyr::spark_connect(master = "local")
  norm <- Normal$new(
    sc = sc,
    size = 10L,
    num_partitions = 1L,
    seed = 1L
  )
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collect"></a>}}
\if{latex}{\out{\hypertarget{method-collect}{}}}
\subsection{Method \code{collect()}}{
Collect the random deviates into R.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$collect()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  norm$collect()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count"></a>}}
\if{latex}{\out{\hypertarget{method-count}{}}}
\subsection{Method \code{count()}}{
Collect the random deviates into R.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$count()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  norm$collect()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-first"></a>}}
\if{latex}{\out{\hypertarget{method-first}{}}}
\subsection{Method \code{first()}}{
Count the number of deviates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$first()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  norm$count()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_num_partitions"></a>}}
\if{latex}{\out{\hypertarget{method-get_num_partitions}{}}}
\subsection{Method \code{get_num_partitions()}}{
Get the number of partitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$get_num_partitions()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  norm$get_num_partitions()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
