% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdds.R
\name{Uniform}
\alias{Uniform}
\title{Generate random deviates from a uniform distribution}
\description{
Generate random deviates from a uniform distribution

Generate random deviates from a uniform distribution
}
\examples{

## ------------------------------------------------
## Method `Uniform$new`
## ------------------------------------------------

\dontrun{
  sc <- sparklyr::spark_connect(master = "local")
  unif <- Uniform$new(
    sc = sc,
    size = 10L,
    num_partitions = 1L,
    seed = 1L
  )
}

## ------------------------------------------------
## Method `Uniform$collect`
## ------------------------------------------------

\dontrun{
  unif$collect()
}

## ------------------------------------------------
## Method `Uniform$count`
## ------------------------------------------------

\dontrun{
  unif$collect()
}

## ------------------------------------------------
## Method `Uniform$first`
## ------------------------------------------------

\dontrun{
  unif$count()
}

## ------------------------------------------------
## Method `Uniform$get_num_partitions`
## ------------------------------------------------

\dontrun{
  unif$get_num_partitions()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Uniform$new()}}
\item \href{#method-collect}{\code{Uniform$collect()}}
\item \href{#method-count}{\code{Uniform$count()}}
\item \href{#method-first}{\code{Uniform$first()}}
\item \href{#method-get_num_partitions}{\code{Uniform$get_num_partitions()}}
\item \href{#method-clone}{\code{Uniform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Uniform} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$new(sc, size, num_partitions, seed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc}}{A \code{spark_connection}.}

\item{\code{size}}{\code{integer(1)}. The number of deviates to generate.}

\item{\code{num_partitions}}{\code{integer(1)}. Number of partitions in the RDD
(default: \code{sc.defaultParallelism}).}

\item{\code{seed}}{\code{integer(1)}. The seed to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Uniform} object which is a \code{RDD[Double]} comprised of
i.i.d. samples ~ \code{U(0.0, 1.0)}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  sc <- sparklyr::spark_connect(master = "local")
  unif <- Uniform$new(
    sc = sc,
    size = 10L,
    num_partitions = 1L,
    seed = 1L
  )
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collect"></a>}}
\if{latex}{\out{\hypertarget{method-collect}{}}}
\subsection{Method \code{collect()}}{
Collect the random deviates into R.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$collect()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  unif$collect()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count"></a>}}
\if{latex}{\out{\hypertarget{method-count}{}}}
\subsection{Method \code{count()}}{
Collect the random deviates into R.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$count()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  unif$collect()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-first"></a>}}
\if{latex}{\out{\hypertarget{method-first}{}}}
\subsection{Method \code{first()}}{
Count the number of deviates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$first()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  unif$count()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_num_partitions"></a>}}
\if{latex}{\out{\hypertarget{method-get_num_partitions}{}}}
\subsection{Method \code{get_num_partitions()}}{
Get the number of partitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$get_num_partitions()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  unif$get_num_partitions()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
