% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{UniformGenerator}
\alias{UniformGenerator}
\title{Generate values from a uniform distribution.}
\description{
Generate values from a uniform distribution.

Generate values from a uniform distribution.
}
\examples{

## ------------------------------------------------
## Method `UniformGenerator$new`
## ------------------------------------------------

\dontrun{
  sc <- sparklyr::spark_connect(master = "local")
  unif_gen <- UniformGenerator$new(sc = sc)
}

## ------------------------------------------------
## Method `UniformGenerator$next_value`
## ------------------------------------------------

\dontrun{
  unif_gen$next_value()
}

## ------------------------------------------------
## Method `UniformGenerator$set_seed`
## ------------------------------------------------

\dontrun{
  unif_gen$set_seed(1L)
}
}
\seealso{
\link{Generators}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UniformGenerator$new()}}
\item \href{#method-next_value}{\code{UniformGenerator$next_value()}}
\item \href{#method-set_seed}{\code{UniformGenerator$set_seed()}}
\item \href{#method-clone}{\code{UniformGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{UniformGenerator} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformGenerator$new(sc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc}}{A \code{spark_connection}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{UniformGenerator} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  sc <- sparklyr::spark_connect(master = "local")
  unif_gen <- UniformGenerator$new(sc = sc)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-next_value"></a>}}
\if{latex}{\out{\hypertarget{method-next_value}{}}}
\subsection{Method \code{next_value()}}{
Get the next value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformGenerator$next_value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{numeric(1)} i.i.d.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  unif_gen$next_value()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_seed"></a>}}
\if{latex}{\out{\hypertarget{method-set_seed}{}}}
\subsection{Method \code{set_seed()}}{
Set random seed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformGenerator$set_seed(seed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{\code{integer(1)}. The unit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}, invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  unif_gen$set_seed(1L)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
